requirements.txt
fastapi
uvicorn
httpx
google-generativeai
python-dotenv
tqdm
gitpython
requests
pydantic